+++ src/main/kotlin/team/mozu/dsm/adapter/in/lesson/dto/response/LessonResponse.kt
@@ -10,7 +10,7 @@ data class LessonResponse(
     val name: String,
     val maxInvRound: Int,
     val curInvRound: Int? = null,
-    val baseMoney: Long,
+    val baseMoney: Int,
     val lessonNum: String? = null,
     val isStarred: Boolean,
     val isDeleted: Boolean,
